+++ packages/waggle/src/storage/storage.ts
@@ -11,7 +11,7 @@ import PeerId from 'peer-id'
 import { CryptoEngine, setEngine } from 'pkijs'
 import {
   ChannelInfoResponse, DataFromPems, IChannelInfo,
-  IMessage, IMessageThread, IPublicKey, IRepo, IZbayChannel, StorageOptions
+  IMessage, IMessageThread, IPublicKey, IRepo, StorageOptions
 } from '../common/types'
 import { Config } from '../constants'
 import logger from '../logger'
@@ -43,7 +43,7 @@ export class Storage {
   public peerId: PeerId
   protected ipfs: IPFS.IPFS
   protected orbitdb: OrbitDB
-  private channels: KeyValueStore<IZbayChannel>
+  private channels: KeyValueStore<IChannelInfo>
   private directMessagesUsers: KeyValueStore<IPublicKey>
   private messageThreads: KeyValueStore<IMessageThread>
   private certificates: EventStore<string>
@@ -53,10 +53,12 @@ export class Storage {
   public options: StorageOptions
   public orbitDbDir: string
   public ipfsRepoPath: string
+  private readonly communityId: string
 
-  constructor(zbayDir: string, io: any, options?: Partial<StorageOptions>) {
+  constructor(zbayDir: string, io: any, communityId: string, options?: Partial<StorageOptions>) {
     this.zbayDir = zbayDir
     this.io = io
+    this.communityId = communityId
     this.options = {
       ...new StorageOptions(),
       ...options
@@ -155,7 +157,7 @@ export class Storage {
 
   private async createDbForChannels() {
     log('createDbForChannels init')
-    this.channels = await this.orbitdb.keyvalue<IZbayChannel>('public-channels', {
+    this.channels = await this.orbitdb.keyvalue<IChannelInfo>('public-channels', {
       accessController: {
         write: ['*']
       }
@@ -165,17 +167,19 @@ export class Storage {
       'replicated',
       // eslint-disable-next-line
       async () => {
-        log('REPLICATED: CHANNELS')
-        if (this.options.isEntryNode) {
-          log('Entry node. Subscribing for all replicated channels')
-          await Promise.all(
-            Object.values(this.channels.all).map(async channel => {
-              if (!this.publicChannelsRepos.has(channel.address)) {
-                await this.subscribeForChannel(channel.address, channel)
-              }
-            })
-          )
-        }
+        log('REPLICATED: Channels')
+        // @ts-expect-error - OrbitDB's type declaration of `load` lacks 'options'
+        await this.channels.load({ fetchEntryTimeout: 2000 })
+        const payload = this.channels.all
+
+        this.io.emit(EventTypesResponse.RESPONSE_GET_PUBLIC_CHANNELS, { communityId: this.communityId, channels: payload })
+        await Promise.all(
+          Object.values(this.channels.all).map(async channel => {
+            if (!this.publicChannelsRepos.has(channel.address)) {
+              await this.subscribeForChannel(channel.address, channel)
+            }
+          })
+        )
       }
     )
 
@@ -270,7 +274,6 @@ export class Storage {
           description: channel.description,
           owner: channel.owner,
           timestamp: channel.timestamp,
-          keys: channel.keys,
           name: channel.name
         }
       }
@@ -398,7 +401,6 @@ export class Storage {
     const channel = this.channels.get(channelAddress)
     if (!channel) {
       await this.channels.put(channelAddress, {
-        orbitAddress: `/orbitdb/${db.address.root}/${db.address.path}`,
         address: channelAddress,
         ...channelData
       })